requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
marshmallow==3.20.1
matplotlib==3.8.2
numpy==1.26.2
pymongo==4.6.1
python-dotenv==1.0.0
scikit_learn==1.3.2
